requirements.txt
telethon==1.34.0
python-telegram-bot==21.10
python-dotenv==1.0.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pytz==2024.1
